DJ_api/requirements.txt
Django==5.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.2
python-dotenv==1.0.0
